Add the ability to run focused tests with karma
Setting the TEST_PATH and TEST_LINE environment variables when running
a karma command (either through `npm test` or `karma` directly) should
try to run just the test defined at that line.

If the line is inside a `describe` rather than an `it`, all tests inside the
`describe` will run.

This is not really useful for the command line, but hopefully useful if
you can convince your editor to invoke karma with these arguments.

The exact mechanisms employed are:
* A helper called test-filter.js is included in the test suite,
  which conditionally sets up jasmine's specFilter
* A preprocessor defined in the karma config determines
  if TEST_PATH and TEST_LINE are used, and if so rewrites
  test-filter.js, telling it to read the focused test name
  from a network request to /focused_test_name.txt
* A middleware defined in the karma config services
  the request to /focused_test_name.txt by reading
  the file at TEST_PATH and building up the full test name
  for the test located at TEST_LINE
* test-filter.js sets up the specFilter so that only
  tests matching that prefix are run.